Place middleware in more appropriately named folders, rename test to match
