+++ wp-content/plugins/core/src/Block_Middleware/Block_Middleware_Definer.php
@@ -13,8 +13,8 @@ use Tribe\Project\Block_Middleware\Pipelines\Add_Fields_Pipeline;
 use Tribe\Project\Block_Middleware\Pipelines\Add_Model_Data_Pipeline;
 use Tribe\Project\Blocks\Blocks_Definer;
 use Tribe\Project\Blocks\Contracts\Model;
-use Tribe\Project\Blocks\Middleware\Color_Theme\Fields\Color_Theme_Field_Middleware;
-use Tribe\Project\Blocks\Middleware\Color_Theme\Models\Color_Theme_Field_Model_Middleware;
+use Tribe\Project\Blocks\Middleware\Color_Theme\Field_Middleware\Color_Theme_Field_Middleware;
+use Tribe\Project\Blocks\Middleware\Color_Theme\Model_Middleware\Color_Theme_Field_Model_Middleware;
 use Tribe\Project\Blocks\Types\Accordion\Accordion;
 use Tribe\Project\Blocks\Types\Accordion\Accordion_Model;
 use Tribe\Project\Blocks\Types\Card_Grid\Card_Grid;
